+++ packages/lib/index.ts
@@ -1,2 +1,3 @@
+export { getLuckyUsers } from "./getLuckyUsers";
 export { default as isPrismaObj, isPrismaObjOrUndefined } from "./isPrismaObj";
 export * from "./isRecurringEvent";
